requirements.txt
httpx==0.27.0
psycopg2-binary==2.9.9
tenacity==8.2.2
python-dotenv==1.0.0
streamlit==1.32.2
pandas==2.2.1
